refactor: move ProductCard to product directory and update imports
