+++ components/product/ProductImage.tsx
@@ -1,9 +1,10 @@
 import React from "react"
+
 import { Box } from "../ui/box"
 import { Image } from "../ui/image"
-import { Product } from "../category/ProductCard"
+import { Product } from "./ProductCard"
 
-const ProductImage = ({product}:{product:Product}) => {
+const ProductImage = ({ product }: { product: Product }) => {
   return (
     <Box className="shadow-lg rounded-full -mt-16 p-2 bg-background-0">
       <Image
